Consolidate npm workspace and update build paths
- Remove separate backend/frontend install and Docker entrypoints
- Point tests, CI, and docs at workspace-aware paths
- Trim dependabot config and root ignore rules
