+++ .github/dependabot.yml
@@ -18,46 +18,12 @@ updates:
         patterns:
           - "*"
 
-  - package-ecosystem: "npm"
-    directory: "/backend"
-    schedule:
-      interval: "weekly"
-      day: "monday"
-      time: "06:15"
-      timezone: "UTC"
-    open-pull-requests-limit: 3
-    groups:
-      backend-production:
-        dependency-type: "production"
-      backend-tooling:
-        dependency-type: "development"
-      backend-transitive:
-        patterns:
-          - "*"
-
-  - package-ecosystem: "npm"
-    directory: "/frontend"
-    schedule:
-      interval: "weekly"
-      day: "monday"
-      time: "06:30"
-      timezone: "UTC"
-    open-pull-requests-limit: 4
-    groups:
-      frontend-production:
-        dependency-type: "production"
-      frontend-tooling:
-        dependency-type: "development"
-      frontend-transitive:
-        patterns:
-          - "*"
-
   - package-ecosystem: "github-actions"
     directory: "/"
     schedule:
       interval: "weekly"
       day: "monday"
-      time: "06:45"
+      time: "06:15"
       timezone: "UTC"
     open-pull-requests-limit: 2
     groups:
@@ -75,31 +41,5 @@ updates:
     open-pull-requests-limit: 1
     groups:
       root-docker:
-        patterns:
-          - "*"
-
-  - package-ecosystem: "docker"
-    directory: "/backend"
-    schedule:
-      interval: "monthly"
-      day: "monday"
-      time: "07:15"
-      timezone: "UTC"
-    open-pull-requests-limit: 1
-    groups:
-      backend-docker:
-        patterns:
-          - "*"
-
-  - package-ecosystem: "docker"
-    directory: "/frontend"
-    schedule:
-      interval: "monthly"
-      day: "monday"
-      time: "07:30"
-      timezone: "UTC"
-    open-pull-requests-limit: 1
-    groups:
-      frontend-docker:
         patterns:
           - "*"
