+++ .github/workflows/validate.yml
@@ -18,18 +18,9 @@ jobs:
         with:
           node-version: 22
           cache: npm
-          cache-dependency-path: |
-            package-lock.json
-            backend/package-lock.json
-            frontend/package-lock.json
+          cache-dependency-path: package-lock.json
 
       - run: npm ci
-
-      - run: npm ci
-        working-directory: backend
-
-      - run: npm ci
-        working-directory: frontend
 
       - run: npm run lint
         working-directory: frontend
