+++ .gitignore
@@ -1,17 +1,38 @@
 node_modules
 npm-debug.log
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
 .env
 .env.local
+.env.development.local
+.env.test.local
+.env.production.local
+logs/
 *.log
 .DS_Store
 frontend/node_modules
 frontend/dist
-docs/
 data/
 backend/data/db.json
 backend/data/*
 !backend/data/.gitkeep
+pids/
+*.pid
+*.seed
+*.pid.lock
+coverage/
+.nyc_output/
+.npm
+.eslintcache
+.node_repl_history
+*.tgz
+.yarn-integrity
+.vscode/
+.idea/
 *.swp
 *.swo
 *~
+*.tmp
+*.temp
 /*.mp3
