+++ .tests/auth/api-auth.integration.test.js
@@ -17,7 +17,7 @@ applyIsolatedBackendEnv(isolatedState);
 const [{ db }, { userOps, dbOps }, bcryptModule] = await Promise.all([
   importFromRepo("backend/config/db-sqlite.js"),
   importFromRepo("backend/config/db-helpers.js"),
-  importFromRepo("backend/node_modules/bcrypt/bcrypt.js"),
+  import("bcrypt"),
 ]);
 
 const bcrypt = bcryptModule.default;
