+++ .tests/auth/lidarr-user-preferences.test.js
@@ -15,7 +15,7 @@ applyIsolatedBackendEnv(isolatedState);
 const [{ db }, { userOps }, bcryptModule] = await Promise.all([
   importFromRepo("backend/config/db-sqlite.js"),
   importFromRepo("backend/config/db-helpers.js"),
-  importFromRepo("backend/node_modules/bcrypt/bcrypt.js"),
+  import("bcrypt"),
 ]);
 
 const bcrypt = bcryptModule.default;
