--- backend/.dockerignore
@@ -1,12 +1,0 @@
-node_modules
-npm-debug.log*
-.env
-.env.local
-.git
-.gitignore
-data/db.json
-README.md
-Dockerfile
-.dockerignore
-.vscode
-.idea
