--- backend/.gitignore
@@ -1,33 +1,0 @@
-node_modules/
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-.env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-logs/
-*.log
-data/db.json
-pids/
-*.pid
-*.seed
-*.pid.lock
-coverage/
-.nyc_output/
-.npm
-.eslintcache
-.node_repl_history
-*.tgz
-.yarn-integrity
-.DS_Store
-Thumbs.db
-.vscode/
-.idea/
-*.swp
-*.swo
-*~
-*.tmp
-*.temp
