--- backend/Dockerfile
@@ -1,27 +1,0 @@
-FROM node:25-alpine
-
-WORKDIR /app
-
-RUN apk add --no-cache su-exec python3 make g++ fontconfig ttf-dejavu && ln -sf /usr/bin/python3 /usr/bin/python
-
-COPY package*.json ./
-
-RUN npm ci --omit=dev --ignore-scripts
-RUN npm rebuild --build-from-source
-
-COPY . .
-COPY docker-entrypoint.sh /usr/local/bin/
-
-RUN chmod +x /usr/local/bin/docker-entrypoint.sh && \
-    addgroup -g 1001 -S nodejs && \
-    adduser -S nodejs -u 1001 && \
-    mkdir -p /app/backend/data && \
-    chown -R nodejs:nodejs /app
-
-EXPOSE 3001
-
-HEALTHCHECK --interval=30s --timeout=10s --start-period=40s --retries=3 \
-  CMD node -e "require('http').get('http://127.0.0.1:3001/api/health', (r) => {process.exit(r.statusCode === 200 ? 0 : 1)})"
-
-ENTRYPOINT ["docker-entrypoint.sh"]
-CMD ["node", "server.js"]
