+++ backend/package.json
@@ -1,7 +1,8 @@
 {
   "name": "aurral-backend",
-  "version": "1.0.0",
+  "version": "1.76.0",
   "description": "Backend API for Aurral artist request manager",
+  "private": true,
   "main": "server.js",
   "type": "module",
   "scripts": {
@@ -32,15 +33,10 @@
     "express-basic-auth": "^1.2.1",
     "express-rate-limit": "^8.5.2",
     "helmet": "^8.2.0",
-    "lowdb": "^7.0.1",
     "music-metadata": "^11.12.3",
     "node-cache": "^5.1.2",
     "sharp": "^0.34.5",
     "slsk-client": "^1.1.0",
-    "uuid": "^14.0.0",
     "ws": "^8.21.0"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.14"
   }
 }
