--- frontend/.dockerignore
@@ -1,13 +1,0 @@
-node_modules
-dist
-npm-debug.log*
-.env
-.env.local
-.git
-.gitignore
-README.md
-Dockerfile
-.dockerignore
-.vscode
-.idea
-.continue
