--- frontend/Dockerfile
@@ -1,29 +1,0 @@
-FROM node:25-alpine AS builder
-
-WORKDIR /app
-
-COPY package*.json ./
-
-RUN npm ci
-
-COPY . .
-
-ARG APP_VERSION=unknown
-ARG GITHUB_REPO=lklynet/aurral
-ENV VITE_APP_VERSION=$APP_VERSION
-ENV VITE_GITHUB_REPO=$GITHUB_REPO
-
-RUN npm run build
-
-FROM nginx:alpine
-
-COPY nginx.conf /etc/nginx/conf.d/default.conf
-
-COPY --from=builder /app/dist /usr/share/nginx/html
-
-EXPOSE 80
-
-HEALTHCHECK --interval=30s --timeout=10s --start-period=40s --retries=3 \
-  CMD wget --quiet --tries=1 --spider http://127.0.0.1:80 || exit 1
-
-CMD ["nginx", "-g", "daemon off;"]
